consumers/requirements.txt
python-dotenv

# Producer
Faker
pendulum
kafka-python
pymongo